mobile: Deflake AndroidEngineExplicitFlowTest (#34252)
This PR fixes the flakiness issue in AndroidEngineExplicitFlowTest that relies on nondeterministic behavior related to the 0 byte when the end stream is true in Envoy Mobile. See envoyproxy/envoy-mobile#1393. The PR also fixes the assertions to first check if there's an Envoy error and fails it other than executing the assertions and fails with NPE when there's an Envoy error.

Risk Level: low (tests only)
Testing: --runs_per_test=100
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

Signed-off-by: Fredy Wijaya <fredyw@google.com>