dfp: adding an experimental option for E-M (#34282)
This adds an off-by-default flag that allows the DFP filter to force resolution if there are no upstream connections. This doesn't work for strict DNS, and is largely intended for use for Envoy mobile. As the TODO states, if the performance is net positive we'll add configuration knobs, otherwise it'll be torn out.

Risk Level: low (off by default)
Testing: new e2e tests
Docs Changes: n/a
Release Notes: n/a
[Optional Runtime guard:] yes

Signed-off-by: Alyssa Wilk <alyssar@chromium.org>