io_handle: cache envoy address instances for recvmsg and recvmmsg (#34649)
Maintain a LRU cache of Address::Instance objects in IoSocketHandleImpl to keep track of source and destination addresses received in recvmsg and recvmmsg. This cache is added mainly used for UDP read. This is an alternative to #33917 to address the issue that each recvmsg/recvmmsg will create a new instance of the source/destination addresses (which does stringification upon instantiation) even though they have been received before.

The cache can be initialized during IoSocketHandleImpl construction based on a new param max_addresses_cache_size_ in SocketCreationOptions. This change only enables EnvoyQuicClientConnection to use a SocketCreationOptions with size 4 to create the underlying IoSocketHandleImpl. There is no change to the server side recvmsg and recvmmsg calls as the significant CPU cost on server side hasn't been observed.

There is no cache support in non-IoSocketHandleImpl-based implementations.

Signed-off-by: Dan Zhang <danzh@google.com>