Allow uri_template variables in literal expression
UriTemplateMatch now allows for variables to be defined in a literal.
Before this change, all variables had to be clamped by path separators (/).

This is not according [RFC6570](https://www.rfc-editor.org/rfc/rfc6570.html).

Even though the spec is not implemented by Envoy (`*`, `**` are not a glob in
the file system sense), it is good to have more feature parity with URI Templates.

The change is to treat path separators as literals.
The restriction can then be easily lifted to allow variables inside literals.

2 checks have been rewritten, double slashes are not allowed, `/abc//d` and
`/abc/{var=/d}`, and that a PathGlob can only exist between two path separators.

Signed-off-by: Jens Kat <jenskat@gmail.com>