mobile: Probe IPv6 connectivity (#34846)
This PR adds a feature to set the IP version to be filtered out from the
DNS cache during the DNS resolution. This feature will be used by Envoy
Mobile to check for the IPv6 connectivity. The check is implemented by
making a `connect` call to Google DNS IPv6 address similar to how it was
implemented in Chromium. On a network change event, the following will
happen:
1. Check if there is an IPv6 connectivity. If not, set a flag to remove
the IPv6 addresses in the DNS cache, so that the next DNS resolution
will filter out the IPv6 addresses from the DNS cache.
2. Force refresh the DNS cache.

This change is needed because the HTTP/3 implementation doesn't
currently implement happy eyeballs, which means if the connection to
IPv6 fails, it will fallback to HTTP/1 or HTTP/2 resulting in a low
HTTP/3 usage.

Risk Level: low
Testing: unit and integration tests
Platform Specific Features: dns, mobile
Runtime guard:
envoy.reloadable_features.dns_cache_set_ip_version_to_remove

---------

Signed-off-by: Fredy Wijaya <fredyw@google.com>