xff: xff_trusted_cidrs should take precedence
This is a revert of d976354a8906b186d4b01f2ab995594b2cca5933 with a
change where Envoy will now log at warn level when both
xff_trusted_cidrs and xff_num_trusted_hops are set as it makes no sense
to do so. This is needed as we can't throw exceptions here due to
exceptions not being allowed in Envoy Mobile.

Signed-off-by: James O'Gorman <james@netinertia.co.uk>