route: move virtualHost() to Route from RouteEntry (#35168)
Move the virtualHost() method to the Route from RouteEntry. By this way,
the filters could access the virtual host even it has no valid
routeEntry().

Move the virtualCluster() to the VirtualHost from RouteEntry because
similar reason.

For example, in pevious implementation, when the users want to log the
metadata of vh by `%METADATA(VIRTUAL_HOST)%`, if direct response is used
in the matched route, then users cannot access the metadata of vh. This
PR make it's possbile.

Risk Level: low.
Testing: n/a.
Docs Changes: n/a.
Release Notes: n/a.
Platform Specific Features: n/a.

Signed-off-by: wbpcode <wbphub@gmail.com>