mobile: Add an option to set the UDP socket send and receive buffer sizes (#35190)
This PR adds an option in the `EngineBuilder` to set the UDP socket send
buffer size. By default it will use the same value as the one set in
Chromium:
https://source.chromium.org/chromium/chromium/src/+/main:net/quic/quic_session_pool.cc;l=790-793;drc=7f04a8e033c23dede6beae129cd212e6d4473d72

This PR also updates the code to only apply the default receive buffer
size to UDP and not TCP given that the code in
https://source.chromium.org/chromium/chromium/src/+/main:net/quic/quic_session_pool.cc;l=766;drc=7f04a8e033c23dede6beae129cd212e6d4473d72
only applies the socket option to UDP only.

Risk Level: low
Testing: unit tests
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: mobile

---------

Signed-off-by: Fredy Wijaya <fredyw@google.com>
Signed-off-by: asingh-g <abhisinghx@google.com>