Create new field for on-demand DNS resolution
This is the first step in providing a sort of on-demand DNS resolution
for Envoy. I chose a cluster-level enum field over changes in
dns_resolution configuration because the enqueueing behavior
of the async DNS requests seem to be a property of the cluster and not
any specific DNS resolver (e.g. c-ares).

Signed-off-by: Keith Mattix II <keithmattix@microsoft.com>