mobile: Add an optional idempotency parameter in sendHeaders (#35474)
This PR adds an optional parameter to indicate whether a request is an
idempotent in the `sendHeaders`. This can be useful to automatically add
a retry policy, such as `http3-post-connect-failure` when the
`idempotent` flag is set. This makes the Envoy Mobile API similar to
Cronet API. The Cronvoy implementation has also been updated to respect
the `idempotency` flag.

Risk Level: low
Testing: unit test
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: mobile

---------

Signed-off-by: Fredy Wijaya <fredyw@google.com>