Refactor Upstream API for setting TCP tunneling mode (#35510)
This API was introduced by #32991 and it conflates support for
half-close with TCP proxying mode in H/1 codec. This interferes with
#34461 which enables support for server half-close in H/2 proxying.

This PR refactors API to separate enabling of the half-close behavior
from TCP proxy behavior. Prior to #34461 half-close semativs were only
enabled for TCP proxy upstreams, but #34461 introduces other cases as
well. This necessitates setting half-close and TCP proxy modes
separately. TCP proxy mode is now enabled in the class dedicated to the
TCP proxy upstreams.

---------

Signed-off-by: Yan Avlasov <yavlasov@google.com>
Signed-off-by: asingh-g <abhisinghx@google.com>