add http1-safe max_connection_duration for http connection manager (#35209)
This PR adds an option to change the http connection manager's draining
behavior for max_connection_duration for http1. The new behavior is that
envoy will wait indefinitely for one more request, add connection:close
to the response headers, then close the connection once the stream ends.

This is to avoid a networking race condition which results in the client
not receiving a response to their last request if they send it right
when envoy is closing the connection.

Fixes #34356 (check for context)

---------

Signed-off-by: antoniovleonti <leonti@google.com>
Signed-off-by: sunaydagli <sunaydagli@google.com>