mobile: Rewrite JNI cache implementation (#35710)
This PR rewrites the JNI cache implementation to fix the GlobalRef leak.
Prior to this PR, some code would call `getObjectClass` in order to get
a `jclass` instance. If the `jclass` was not already in the cache, the
code would create a new `GlobalRef` and add it into the cache. Given
that there's no guarantee that we will get the same `jclass` instance,
the previous code ended up creating a lot of `new GlobalRef` instances
due to the cache misses causing a leak. The new implementation updates
the code to load the `jclass` from the cache directly via
`findClassFromCache` instead of using `getObjectClass`. This guarantees
that we will use the same `jclass` instance. The new implementation also
changes the cache initialization to be eager instead of lazy, thereby
eliminating the need for having `thread_local`, which can help to
further reduce the number of `GlobalRef` instances.

Risk Level: medium
Testing: unit tests
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: mobile

Signed-off-by: Fredy Wijaya <fredyw@google.com>