ext_proc: expose main stream info from the async client steam info (#35650)
Some upstream HTTP filters running in the ext_proc sidestream may
require access to metadata associated with the main stream established
for the client. This change exposes the main ("parent") stream_info from
the async client's stream info. We need to clear this (to avoid a
dangling reference) in the case of deferred stream closure, where the
main stream closes before the async client stream.

---------

Signed-off-by: Elisha Ziskind <eziskind@google.com>