http: modify authority validation to allow @ character (#35602)
Envoy uses the `authorityIsValid()` to validate the authority header for
both H/1 and H/2 codecs.
Previously Envoy used the nghttp2 validator and in #24943 this was
changed to oghttp2's implementation.
The two implementations differ in the way they handle the "@" character
(nghttp2 allows it, and oghttp2 doesn't).
According to the H/2 spec, the "@" character is not allowed as part of
the authority header. However, for H/1 it is allowed as part of the
"user-info@host:port" structure of the authority header.
This PR changes the validator to be similar to the nghttp2
implemenation.
The change can be temporarily dis

In the future, when Envoy fully supports UHV (#10646), the H/1 and H/2
validation parts should be decoupled, and the oghttp2 authority
validation can be used for H/2.

---------

Signed-off-by: Adi Suissa-Peleg <adip@google.com>
Signed-off-by: sunaydagli <sunaydagli@google.com>