xds-failover: add stat tracking primary/failover connection (#35691)
Extends the possible values of the `control_plane.connected_state`
gauge. Previously it only stored 0 when Envoy wasn't
connecting/connected to the xDS server, and 1 when it was.
This PR changes it to have the following values:
0 - Envoy isn't connected/connecting to an xDS server.
1 - Envoy is connecting/connected to the primary source.
2 - Envoy is connecting/connected to the failover source.

Risk Level: low.
Testing: Added integration tests.
Docs Changes: Will be added once the feature is complete
Release Notes: N/A.
Platform Specific Features: N/A

Signed-off-by: Adi Suissa-Peleg <adip@google.com>
Signed-off-by: sunaydagli <sunaydagli@google.com>