[cache_filter] Determine end_stream later (#35699)
Commit Message: [cache_filter] Determine end_stream later
Additional Description: Currently the cache filter is incompatible with
streaming a result at the same time as populating it, because the lookup
requires foreknowledge of whether the entry will have trailers or not,
which isn't known until the end of the request (a `Trailers` header is a
hint but not a guarantee). This change makes it so the end_stream value
from the cache acts more like the way it does for every other stream, so
no foreknowledge is required. Plus similar changes for content-length
being potentially unknown under the same circumstances.
Risk Level: Some risk to changing behavior of a filter still tagged WIP.
Testing: Existing tests provide coverage, and are updated to verify the
new value. Additional tests for content-length handling especially
around range requests.
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

---------

Signed-off-by: Raven Black <ravenblack@dropbox.com>
Signed-off-by: sunaydagli <sunaydagli@google.com>