mobile: Update Android proxy tests to not use mocks (#35849)
This PR updates the Android proxy tests to not use mocks, which have
been causing an issue for me when running the tests locally due to the
bytecode mismatch. Using a real object with `shadowOf` is also the
recommended approach. This PR also updates the tests to not use a
deprecated `sendStickyBroadcast` method.

Risk Level: low (tests only)
Testing: unit tests
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

Signed-off-by: Fredy Wijaya <fredyw@google.com>