quic: Connect client UDP sockets before usage (#35652)
This change does a few things:
1. Calls `connect()` on the QUIC UDP sockets, which allows the
connection to receive async ICMP error messages on `recvmsg` calls
throughout the lifetime of the connection.
2. Does not call `bind()` on a QUIC client UDP socket unless the local
address is specified.
3. If the socket is already connected, call `writev`/`send` instead of
`sendmsg`. Connected sockets should not specify a peer address and
hence, writev/send is more appropriate.
4. Previously, when `EnvoyQuicClientConnection` did preferred server
address probing, it specified the existing remote address, not the new
preferred server address. This still worked because the `sendmsg` call
took in the actual destination IP address. However, when calling
`connect()`, this caused the socket to connect to the wrong address.
This issue is fixed in this PR by creating a socket connection to the
preferred server address.
5. Uses the loopback address for the local address, if one isn't
specified, as it is only needed in `createConnectionSocket` to get an
`IoHandle`. This avoids having to make the more expensive `getifaddrs`
call, as mentioned in https://github.com/envoyproxy/envoy/issues/35137.
6. The `connect()` behavior is guarded by the runtime guard
`envoy.reloadable_features.quic_connect_client_udp_sockets`, defaulted
to `true`.

Fixes #35137

---------

Signed-off-by: Ali Beyad <abeyad@google.com>