ext_proc: fix an issue with trailer (#35923)
If ext_proc filter is configured to send header, body but not send
trailer, if the downstream request has trailer after body, then there is
a problem when handling body response.

It supposes to return continueDecoding() to let the filter manager to
send the trailer and finish the request sending. However, the current
logic is missing that. This ends up timeout in this situation, and local
reply is sent to downstream client.

Risk Level: low
Testing: inline
Docs Changes: N/A

Signed-off-by: Yanjun Xiang <yanjunxiang@google.com>