Make http3 codec not encode empty trailers block (#35907)
Commit Message: Make http3 codec not encode empty trailers block
Additional Description: http3 and `grpc_web_filter` combined resulted in
an incompatibility with Chrome - a grpc web request received a response
with an empty trailers block, which Chrome treats as a network error.
The http2 codec had a special case in it for empty trailers, which was
not mimicked in the http3 codec. This PR makes the http3 codec similar
to the http2 codec in this respect, and adds an integration test
ensuring that this behavior is in all downstream codecs.
Risk Level: Small. It is a behavior change, but from an already
malfunctioning state in a rare edge case.
Testing: Added integration test which fails before the change and passes
after.
Docs Changes: n/a
Release Notes: Yes
Platform Specific Features: n/a

---------

Signed-off-by: Raven Black <ravenblack@dropbox.com>