quic: enable certificate compression/decompression
QUIC servers are limited in how much data they can send to clients
in the ServerHello before the client is validated (see
https://www.rfc-editor.org/rfc/rfc9000.html#section-8). If too much
data needs to be sent, an extra network round trip is needed.

One way to reduce the size of the ServerHello data is to compress the
certificates. This can, in some situations, remove an extra round trip.

Signed-off-by: Greg Greenway <ggreenway@apple.com>