json: replace literal string to constexpr and force Streamer to use simple output abstraction (#35892)
Continuous work of #35891. This replace the literal string to constexpr
string view to make it's more friendly for dev tools. And this PR also
create a simple wrapper to the Buffer::Instance to force the Streamer
only access the limited abstraction (only provides an add()) to
streaming it's output.

Risk Level: low.
Testing: n/a.
Docs Changes: n/a.
Release Notes: n/a.
Platform Specific Features: n/a.

---------

Signed-off-by: wangbaiping <wbphub@gmail.com>