http: fix issues in CONNECT-UDP forwarding mode.
When Envoy operates as a CONNECT-UDP forwarding proxy, it was resetting
the upstream stream because it received HTTP Datagrams before receiving
the SETTINGS frame. A new enum has been added in QUICHE to distinguish
this case, so I added handling logic for this and made Envoy drop the
datagrams instead of resetting the stream.

Also, Envoy was dropping Datagrams because the default maximum packet
length for QUIC connections in QUICHE is not large enough for tunneling
use cases such as CONNECT-UDP. I made the cluster manager override the
maximum packet length for upstream QUIC connections to fix this issue,
but only when the downstream request is CONNECT-UDP.

Signed-off-by: Jeongseok Son <jeongseok.son@gmail.com>