Refactor async files library to use dispatcher for synchronization (#36063)
Using mutexes in the async file library was a
premature optimization that made it significantly more complicated and
error-prone to use from the client side. The benefit was the ability to
perform multiple file operations sequentially without handing control
back to the calling thread, but the cost in complexity was not worth it.
This change is a hefty simplification, the main benefit of which is, in
this model, cancellation is simple - previously calling cancel couldn't
guarantee anything so the client object would still have to handle the
case of a callback potentially being called after the object's own
destruction. In this simplified model, calling cancel from the
originating thread, before the callback has been called, guarantees that
the callback will not be called.

Similar simplification may be possible in `cache_filter`, but it's not
quite as clear; cache_filter currently uses `dispatcher->post` to ensure
thread locality, which is now guaranteed by the cache API so that usage
is unnecessary, but it also uses `dispatcher->post` to ensure that
callbacks are not called during the current function context (so that,
e.g. `continueDecoding()` isn't potentially called before
`decodeHeaders` returns `StopAllIterationAndWatermark`, as would happen
with the simple in-memory cache which calls the callback immediately).
Since this is already a giant PR, and it already works as a
simplification with `cache_filter` unchanged, I'm electing to defer a
simplifying pass over `cache_filter` to a followup PR.

Risk Level: Small, only WIP filters use the library, and all existing
tests (with appropriate modifications) still pass - the only test
removed was one testing the "not handing control back to the caller"
sequential action behavior, which now doesn't exist. (And there are
apparently some crash-bugs in the existing stuff, so there's some
negative risk in that this might fix a bug!)
Testing: Existing coverage.
Docs Changes: Yes, code-only.
Release Notes: n/a
Platform Specific Features: n/a

---------

Signed-off-by: Raven Black <ravenblack@dropbox.com>