MockDispatcher to not accidentally run on other threads (#36134)
Commit Message: MockDispatcher to not accidentally run on other threads
Additional Description: Several issues have been found recently where
the MockDispatcher's default behavior of just running the callback
immediately causes tests to pass, but asan/tsan/release tests fail,
because it ends up running the callback on a different thread, exactly
the opposite of almost the whole point of posting things on a dispatcher
which is to put them on a specific thread. This change makes
MockDispatcher insist on being on whichever thread created the
dispatcher, most likely the main thread in tests. This is still not
great because the behavior of running the callback immediately is still
quite different from the behavior of a dispatcher in production, e.g.
```
auto x = std::make_shared<int>(0);
this_thread_dispatcher.post([x]() { *x += 5; });
*x += 1;
std::cerr << *x << std::endl;
```
would always output 6 with a default MockDispatcher, and would always
output 1 with a production dispatcher, but at least cases like that can
be debugged linearly and will be the same in all test environments,
versus the running on a different (and incorrect) thread problem is very
difficult to debug and sometimes only shows up (and sometimes flakily)
in the harder to run test contexts.
(But still pretty bad in that test coverage would say a thing works,
then in production it could do a different thing than it did in tests.)

Since it's a mock the behavior can also be manually overridden to be the
old way. Arguably it should have just always required explicit manual
interception because it's supposed to be a mock not a fake.

Risk Level: Should be safe since it's test-only.
Testing: This is test-only.
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

---------

Signed-off-by: Raven Black <ravenblack@dropbox.com>