Remove some Undefined Behavior
It's UB to specialize `std::ostream`'s `operator<<` on `std::vector`
and `std::pair`.

Replace uses of it in tests with GoogleTest's `::testing::PrintToString()`

While we're at it, remove an unnecessariy opening of namespace std for
specializing `std::hash` on `Envoy::HashedValue`. This is okay, but we can
do it without opening the namespace.

Signed-off-by: Alejandro R. Sedeño <asedeno@google.com>