Use ECDSA capabilities of client to inform TlsContext choice
We now use the ECDSA capabilities of the client to inform certificate choice
The `isClientEcdaCapable` method is now `getClientEcdsaCapabilities`, returning
an optional vector of curves supported by the client. We choose a context
based on whether there is a shared curve in a certificate and the client
capabilities.

Quic helper functions have been change to define the P-256 curve as the client
capability used where they had previously only asserted `true`. Given that QUICHE
only supports P-256 and P-384 certs, no further effort on QUIC has been expended
as part of this PR.

Signed-off-by: anitabyte <anita@anitabyte.xyz>