[cache_filter] Separate upstream from downstream (#36309)
Commit Message: [cache_filter] Separate upstream from downstream
Additional Description: A recurring issue with cache_filter is that
because it uses the regular upstream from a client's connection to
populate the cache, a client being slow to download, or disconnecting
mid-stream, can cause the cache-populating upstream to be delayed by
watermarks or aborted. If the cache were to attempt to stream the
results while they're still being populated, this can lead to many
clients all receiving an unexpected disconnect or an unnecessarily slow
response.
This PR makes the cache filter use a separate upstream made with
AsyncHttpClient, so even if the client disconnects the upstream can
persist and continue to populate the cache entry.

There are two flaws in this version.
1. the downstream write and the cache write are still being performed in
parallel so if one is slower than the other the slower one's buffer can
become oversized in memory
2. the upstream read has no speed limit so if the cache write is slower
than the network read then the cache write buffer can become oversized
in memory.

Subsequent PRs are intended to address these:
1. once the cache provides a "streamed" interface from the cache while
the entry is being written, we can use that for the "primary" downstream
the same as any other downstream requesting the same resource, rather
than streaming from the "shared read" as it does now.
2. adding an API for the ability to pause and resume an AsyncHttpClient
Stream's "read" interface so if the cache-write is a bottleneck the
amount of data buffered in memory is capped.
3. cache filter also needs to monitor downstream watermark callbacks to
avoid pushing data faster than the downstream consumes it.

Risk Level: Low, filter is WIP and previously bug-ridden and
crash-prone. Integration test passes with no changes after this change.
Testing: Mostly covered by existing tests - one test changed from
"expect abort" to "expect cache write completion", as is the intent of
this PR.
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

---------

Signed-off-by: Raven Black <ravenblack@dropbox.com>